pyproject.toml
[tool.poetry]
name = "3t-bot"
version = "0.1.0"
description = "A bot for signing up to pilates reformer and powerpit"
authors = ["Taheera Ahmed <taheera.ahmed@nina.no>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
playwright = "^1.50.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
